refactor: sort folders basket details shop components
