+++ src/components/header/Icons-stack.tsx
@@ -5,7 +5,7 @@ import { Badge } from '@mui/material';
 import IconButton from '@mui/material/IconButton';
 import { NavLink } from 'react-router';
 
-import { LineItem } from '../../pages/cart/clearCartObject.ts';
+import { LineItem } from '../../pages/basket/utils/clearCartObject.ts';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks.ts';
 import { getCustomer, getThemeName } from '../../redux/selectors.ts';
 import { setTheme } from '../../redux/slices/themeSlice.ts';
