+++ src/interfaces/customerDeteils.interface.ts
