+++ src/pages/details/utils/findDiscount.ts
@@ -1,0 +1,5 @@
+export function findDiscount(price: number, discount: number): number {
+  const discountValue = price * (discount / 10000);
+  const countedDiscount = price - discountValue;
+  return countedDiscount;
+}
