+++ src/pages/details/utils/formatPrice.ts
@@ -10,9 +10,3 @@ export function formatPrice(currency: string, price: number): string {
   }
   return '';
 }
-
-export function findDiscount(price: number, discount: number): number {
-  const discountValue = price * (discount / 10000);
-  const countedDiscount = price - discountValue;
-  return countedDiscount;
-}
