+++ src/redux/slices/asyncThunks/loginCustomer.ts
@@ -2,7 +2,7 @@ import { createAsyncThunk } from '@reduxjs/toolkit';
 const projectKey: string = import.meta.env.VITE_CTP_PROJECT_KEY as string;
 
 import { loginCustomerClient } from '../../../ecommerce/clientBuilder';
-import { ICustomerDetails } from '../../../interfaces';
+import { ICustomerDetails } from '../../../interfaces/customerDeteils.interface';
 import { ELocalStorage } from '../../../services/interfaces/createCart.interface';
 
 export interface ILoginCredentials {
